+++ src/main/java/fll/Launcher.java
@@ -30,6 +30,7 @@ import java.net.Socket;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Arrays;
@@ -68,6 +69,7 @@ import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.SystemUtils;
 import org.checkerframework.checker.initialization.qual.UnknownInitialization;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -87,6 +89,7 @@ import fll.util.FLLRuntimeException;
 import fll.util.FormatterUtils;
 import fll.util.GuiExceptionHandler;
 import fll.web.UserRole;
+import fll.web.WebUtils;
 import fll.web.setup.CreateDB.UserAccount;
 import fll.xml.ChallengeDescription;
 import fll.xml.ui.ChallengeDescriptionEditor;
@@ -646,13 +649,13 @@ public class Launcher extends JFrame {
         try {
           Desktop.getDesktop().open(dir.toFile());
         } catch (final IOException e) {
-          final String message = "Error opening sponsor_logos directory: "
+          final String message = "Error opening sponsor logos directory: "
               + e.getMessage();
           LOGGER.error(message, e);
           JOptionPane.showMessageDialog(this, message, "ERROR", JOptionPane.ERROR_MESSAGE);
         }
       } else {
-        JOptionPane.showMessageDialog(this, "Cannot find sponsor_logos directory.", "ERROR", JOptionPane.ERROR_MESSAGE);
+        JOptionPane.showMessageDialog(this, "Cannot find sponsor logos directory.", "ERROR", JOptionPane.ERROR_MESSAGE);
       }
     });
 
@@ -945,13 +948,13 @@ public class Launcher extends JFrame {
   /**
    * @return the directory or null if not found
    */
-  private @Nullable Path getSponsorLogosDirectory(@UnknownInitialization(Launcher.class) Launcher this) {
+  private static @Nullable Path getSponsorLogosDirectory() {
     final Path classesPath = TomcatLauncher.getClassesPath();
     final Path webroot = TomcatLauncher.findWebappRoot(classesPath);
     if (null == webroot) {
       return null;
     } else {
-      final Path check = webroot.resolve("sponsor_logos");
+      final Path check = webroot.resolve(WebUtils.SPONSOR_LOGOS_PATH);
       if (Files.exists(check)
           && Files.isDirectory(check)) {
         return check.normalize();
@@ -964,13 +967,13 @@ public class Launcher extends JFrame {
   /**
    * @return the directory or null if not found
    */
-  private @Nullable Path getSlideshowDirectory(@UnknownInitialization(Launcher.class) Launcher this) {
+  private static @Nullable Path getSlideshowDirectory() {
     final Path classesPath = TomcatLauncher.getClassesPath();
     final Path webroot = TomcatLauncher.findWebappRoot(classesPath);
     if (null == webroot) {
       return null;
     } else {
-      final Path check = webroot.resolve("slideshow");
+      final Path check = webroot.resolve(WebUtils.SLIDESHOW_PATH);
       if (Files.exists(check)
           && Files.isDirectory(check)) {
         return check.normalize();
@@ -1236,6 +1239,34 @@ public class Launcher extends JFrame {
     } catch (final IOException e) {
       throw new FLLRuntimeException("Error creating temporary file", e);
     }
+
+    final Path oldBaseDir = Paths.get(oldInstallationDirectory);
+    final Path oldWebDir = oldBaseDir.resolve("web");
+
+    // copy sponsor logos
+    final Path oldSponsorLogos = oldWebDir.resolve(WebUtils.SPONSOR_LOGOS_PATH);
+    final Path newSponsorLogos = getSponsorLogosDirectory();
+    if (null == newSponsorLogos) {
+      throw new FLLRuntimeException("Unable to find current sponsor logos directory");
+    }
+    try {
+      FileUtils.copyDirectory(oldSponsorLogos.toFile(), newSponsorLogos.toFile());
+    } catch (final IOException e) {
+      throw new FLLRuntimeException("Error copying sponsor logos", e);
+    }
+
+    // copy slideshow
+    final Path oldSlideshow = oldWebDir.resolve(WebUtils.SLIDESHOW_PATH);
+    final Path newSlideshow = getSlideshowDirectory();
+    if (null == newSlideshow) {
+      throw new FLLRuntimeException("Unable to find current slideshow directory");
+    }
+    try {
+      FileUtils.copyDirectory(oldSlideshow.toFile(), newSlideshow.toFile());
+    } catch (final IOException e) {
+      throw new FLLRuntimeException("Error copying slideshow", e);
+    }
+
   }
 
 }