+++ Changes.md
@@ -1,3 +1,6 @@
+* #1108 - allow one to manipulate slideshow images from the web
+* Copy slideshow and sponsor logs during migration from an older version of the software
+* #1107 - allow one to manipulate sponsor logos from the web
 * Fix bug in schedule UI where the default value of a time is a string
 * #1106 - Ensure no duplicate rows in most recent scores scoreboard display
 * #1105 - Rework dynamic display to properly work with stand alone scoreboard pages